Use a "deferred stat application" approach to handle stat change rollback in incremental undos.
Specifically, do not actually apply stat changes during undoable actions (playPieces), but instead store the pending changes in a new DB table (`turn_progress_stats`), the rows of which ae set to `DELETE CASCADE` on the corresponding move in the `turn_progress` table.  Then, when the turn is "committed" (`donePlayPieces` state), read those changes and apply them. In the case of `undo`, deletion of the primary key row in `turn_progress` throws away pending stat changes.

Note that this means that stats "lag" until end of turn, but intra-game stat access are not useful or typically done (and not surfaced in any player-visible UI).